+++ src/main/java/edu/fiuba/algo3/modelo/Evento/EventoPuntaje/AumentoPorcentaje.java
@@ -3,8 +3,8 @@ package edu.fiuba.algo3.modelo.Evento.EventoPuntaje;
 import edu.fiuba.algo3.modelo.Movimiento.Movimiento;
 
 public class AumentoPorcentaje implements EventoPuntaje {
-    private Integer porcentaje;
-    public AumentoPorcentaje(Integer porcentaje) {
+    private Float porcentaje;
+    public AumentoPorcentaje(Float porcentaje) {
         this.porcentaje = porcentaje;
     }
